+++ client/_component/component.js
@@ -1,12 +1,12 @@
 Template.Mongol_Component.events({
 	'click .Mongol_row': function () {
-		if (Session.equals("Mongol_currentCollection", this.name)) {
-		  Session.set("Mongol_currentCollection", null);
+		if (MeteorToysDict.equals("Mongol_currentCollection", this.name)) {
+		  MeteorToysDict.set("Mongol_currentCollection", null);
 		} else {
-		  Session.set("Mongol_currentCollection", this.name);
+		  MeteorToysDict.set("Mongol_currentCollection", this.name);
 		}
 
-		Session.set("Mongol_editMode", false);
+		MeteorToysDict.set("Mongol_editMode", false);
 	},
 	'click .Mongol_contentView': function (e) {
 		e.stopPropagation();
@@ -15,7 +15,7 @@ Template.Mongol_Component.events({
 
 Template.Mongol_Component.helpers({
   active: function () {
-    if (Session.equals("Mongol_currentCollection", this.name)) {
+    if (MeteorToysDict.equals("Mongol_currentCollection", this.name)) {
       return "Mongol_row_expand";
     }
   }