+++ client/row_collection/collections.js
@@ -4,7 +4,7 @@ Template.Mongol_collection.events({
     var targetCollection = String(this),
         sessionKey       = "Mongol_" + targetCollection;
 
-    if (Session.equals("Mongol_currentCollection", targetCollection)) {
+    if (MeteorToysDict.equals("Mongol_currentCollection", targetCollection)) {
       
       // do nothing
     
@@ -13,8 +13,8 @@ Template.Mongol_collection.events({
       // If the collection doesn't have an index key set,
       // start it from the first document
       
-      if (!Session.get(sessionKey)) {
-        Session.set(sessionKey, 0);
+      if (!MeteorToysDict.get(sessionKey)) {
+        MeteorToysDict.set(sessionKey, 0);
       }
       
     }
@@ -38,7 +38,7 @@ Template.Mongol_collection.helpers({
     var targetCollection = String(this);
     var sessionKey = "Mongol_" + targetCollection;
 
-    var current = Session.get(sessionKey);
+    var current = MeteorToysDict.get(sessionKey);
     var count = current + 1;
 
     return count;