+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name:    'msavin:mongol',
   summary: 'The insanely handy development package for Meteor.',
-  version: '1.0.97',
+  version: '1.0.98',
   git:     'https://github.com/msavin/Mongol.git',
   documentation: 'README.md',
   debugOnly: true
@@ -46,13 +46,15 @@ Package.onUse(function(api) {
 
   api.versionsFrom('1.0');
 
-  api.use(['templating','tracker','mongo','session'], 'client');
+  api.use(['templating','tracker','mongo'], 'client');
+  
   // This must go before: api.use('dburles:mongo-collection-instances@0.3.1');
   // Weak dependency: only used if app contains package 
   api.use('aldeed:collection2@2.3.2', {weak: true}); 
+  
   api.use('dburles:mongo-collection-instances@0.3.3');
   api.use('babrahams:editable-json@0.4.3');
-  api.use('meteortoys:toykit@0.7.0');
+  api.use('meteortoys:toykit@0.8.1');
   
   api.add_files(commonFiles);
   api.add_files(clientFiles, "client");